Removed Desk.js. Consolidated
